Improve assets module structure
