+++ primitives/src/types.rs
@@ -17,7 +17,7 @@
 // along with Zeitgeist. If not, see <https://www.gnu.org/licenses/>.
 
 pub use crate::{
-    asset::*, market::*, max_runtime_usize::*, outcome_report::OutcomeReport, pool::*,
+    assets::*, market::*, max_runtime_usize::*, outcome_report::OutcomeReport, pool::*,
     pool_status::PoolStatus, proxy_type::*,
 };
 #[cfg(feature = "arbitrary")]
